requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
cryptography==42.0.5
apscheduler==3.10.4
httpx==0.27.0
google-api-python-client==2.122.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
librouteros==3.4.0
